Move getSelectedData, getOrderedSelectedMappedKeys, back to internal/actions
